@@ -3,7 +3,7 @@ requires = [ "poetry-core>=1.0.0",]
 build-backend = "poetry.core.masonry.api"
 
 [tool.poetry]
-version = "6.0.0"
+version = "6.1.0"
 name = "source-amazon-ads"
 description = "Source implementation for Amazon Ads."
 authors = [ "Airbyte <contact@airbyte.io>",]
@@ -16,15 +16,14 @@ repository = "https://github.com/airbytehq/airbyte"
 include = "source_amazon_ads"
 
 [tool.poetry.dependencies]
-python = "^3.9,<3.12"
-airbyte-cdk = "0.90.0"
-pendulum = "==2.1.2"
+python = "^3.10,<3.12"
+airbyte-cdk = "^5"
 
 [tool.poetry.scripts]
 source-amazon-ads = "source_amazon_ads.run:run"
 
 [tool.poetry.group.dev.dependencies]
-responses = "^0.23.1"
+responses = "^0.25"
 freezegun = "*"
 requests-mock = "*"
 pytest-mock = "*"
