@@ -7,11 +7,12 @@ import logging
 from typing import Any, List, Mapping, Optional, Tuple
 
 import pendulum
+from airbyte_cdk.models import AdvancedAuth, AuthFlowType, ConnectorSpecification, DestinationSyncMode, OAuthConfigSpecification
 from airbyte_cdk.sources import AbstractSource
 from airbyte_cdk.sources.streams import Stream
 from airbyte_cdk.sources.streams.http.requests_native_auth import Oauth2Authenticator
 
-from .schemas import Profile
+from .spec import SourceAmazonAdsSpec
 from .streams import (
     AttributionReportPerformanceAdgroup,
     AttributionReportPerformanceCampaign,
@@ -132,7 +133,11 @@ class SourceAmazonAds(AbstractSource):
             AttributionReportProducts,
         ]
         portfolios_stream = Portfolios(**stream_args)
-        return [profiles_stream, portfolios_stream, *[stream_class(**stream_args) for stream_class in non_profile_stream_classes]]
+        return [
+            profiles_stream,
+            portfolios_stream,
+            *[stream_class(**stream_args) for stream_class in non_profile_stream_classes],
+        ]
 
     @staticmethod
     def _make_authenticator(config: Mapping[str, Any]):
@@ -144,13 +149,51 @@ class SourceAmazonAds(AbstractSource):
         )
 
     @staticmethod
-    def _choose_profiles(config: Mapping[str, Any], available_profiles: List[Profile]):
+    def _choose_profiles(config: Mapping[str, Any], available_profiles: List[dict[str, Any]]):
         requested_profiles = config.get("profiles", [])
         requested_marketplace_ids = config.get("marketplace_ids", [])
         if requested_profiles or requested_marketplace_ids:
             return [
                 profile
                 for profile in available_profiles
-                if profile.profileId in requested_profiles or profile.accountInfo.marketplaceStringId in requested_marketplace_ids
+                if profile["profileId"] in requested_profiles or profile["accountInfo"]["marketplaceStringId"] in requested_marketplace_ids
             ]
         return available_profiles
+
+    def spec(self, logger: logging.Logger) -> ConnectorSpecification:
+
+        return ConnectorSpecification(
+            documentationUrl="https://docs.airbyte.com/integrations/sources/amazon-ads",
+            connectionSpecification=SourceAmazonAdsSpec.schema(),
+            supportsDBT=False,
+            advanced_auth=AdvancedAuth(
+                auth_flow_type=AuthFlowType.oauth2_0,
+                predicate_key=["auth_type"],
+                predicate_value="oauth2.0",
+                oauth_config_specification=OAuthConfigSpecification(
+                    oauth_user_input_from_connector_config_specification={
+                        "type": "object",
+                        "additionalProperties": False,
+                        "properties": {"region": {"type": "string", "path_in_connector_config": ["region"]}},
+                    },
+                    complete_oauth_output_specification={
+                        "type": "object",
+                        "additionalProperties": True,
+                        "properties": {"refresh_token": {"type": "string", "path_in_connector_config": ["refresh_token"]}},
+                    },
+                    complete_oauth_server_input_specification={
+                        "type": "object",
+                        "additionalProperties": True,
+                        "properties": {"client_id": {"type": "string"}, "client_secret": {"type": "string"}},
+                    },
+                    complete_oauth_server_output_specification={
+                        "type": "object",
+                        "additionalProperties": True,
+                        "properties": {
+                            "client_id": {"type": "string", "path_in_connector_config": ["client_id"]},
+                            "client_secret": {"type": "string", "path_in_connector_config": ["client_secret"]},
+                        },
+                    },
+                ),
+            ),
+        )
