@@ -1,19 +1,18 @@
 #
 # Copyright (c) 2023 Airbyte, Inc., all rights reserved.
 #
-
-from abc import ABC, abstractmethod
+import logging
+from abc import ABC
 from http import HTTPStatus
-from typing import Any, Iterable, List, Mapping, MutableMapping, Optional
+from typing import Any, Iterable, List, Mapping, MutableMapping, Optional, Union
 
 import requests
+from airbyte_cdk.sources.declarative.models import FailureType
 from airbyte_cdk.sources.streams.core import Stream
 from airbyte_cdk.sources.streams.http import HttpStream
-from airbyte_cdk.sources.utils.schema_helpers import expand_refs
-from pydantic import BaseModel, ValidationError
+from airbyte_cdk.sources.streams.http.error_handlers import ErrorHandler, ErrorResolution, HttpStatusErrorHandler, ResponseAction
+from pydantic.v1 import BaseModel, ValidationError
 from source_amazon_ads.constants import URL_MAPPING
-from source_amazon_ads.schemas import CatalogModel
-from source_amazon_ads.schemas.profile import Profile
 
 """
 This class hierarchy may seem overcomplicated so here is a visualization of
@@ -64,6 +63,8 @@ reports for profiles from BasicAmazonAdsStream _profiles list.
 
 """
 
+LOGGER = logging.getLogger("airbyte")
+
 
 class ErrorResponse(BaseModel):
     code: str
@@ -71,27 +72,40 @@ class ErrorResponse(BaseModel):
     requestId: Optional[str]
 
 
+class AmazonAdsErrorHandler(HttpStatusErrorHandler):
+    def interpret_response(self, response_or_exception: Optional[Union[requests.Response, Exception]] = None) -> ErrorResolution:
+
+        if response_or_exception.status_code == HTTPStatus.OK:
+            return ErrorResolution(ResponseAction.SUCCESS)
+
+        try:
+            resp = ErrorResponse.parse_raw(response_or_exception.text)
+        except ValidationError:
+            return ErrorResolution(
+                response_action=ResponseAction.FAIL,
+                failure_type=FailureType.system_error,
+                error_message=f"Response status code: {response_or_exception.status_code}. Unexpected error. {response_or_exception.text=}",
+            )
+
+        LOGGER.warning(
+            f"Unexpected error {resp.code} when processing request {response_or_exception.request.url} for "
+            f"{response_or_exception.request.headers['Amazon-Advertising-API-Scope']} profile: {resp.details}"
+        )
+
+        return ErrorResolution(ResponseAction.SUCCESS)
+
+
 class BasicAmazonAdsStream(Stream, ABC):
     """
     Base class for all Amazon Ads streams.
     """
 
-    def __init__(self, config: Mapping[str, Any], profiles: List[Profile] = None):
+    is_resumable = False
+
+    def __init__(self, config: Mapping[str, Any], profiles: List[dict[str, Any]] = None):
         self._profiles = profiles or []
         self._client_id = config["client_id"]
         self._url = URL_MAPPING[config["region"]]
-
-    @property
-    @abstractmethod
-    def model(self) -> CatalogModel:
-        """
-        Pydantic model to represent json schema
-        """
-
-    def get_json_schema(self):
-        schema = self.model.schema()
-        expand_refs(schema)
-        return schema
 
 
 # Basic full refresh stream
@@ -102,7 +116,7 @@ class AmazonAdsStream(HttpStream, BasicAmazonAdsStream):
 
     data_field = ""
 
-    def __init__(self, config: Mapping[str, Any], *args, profiles: List[Profile] = None, **kwargs):
+    def __init__(self, config: Mapping[str, Any], *args, profiles: List[dict[str, Any]] = None, **kwargs):
         # Each AmazonAdsStream instance are dependant on list of profiles.
         BasicAmazonAdsStream.__init__(self, config, profiles=profiles)
         HttpStream.__init__(self, *args, **kwargs)
@@ -110,10 +124,6 @@ class AmazonAdsStream(HttpStream, BasicAmazonAdsStream):
     @property
     def url_base(self):
         return self._url
-
-    @property
-    def raise_on_http_errors(self):
-        return False
 
     def next_page_token(self, response: requests.Response) -> Optional[Mapping[str, Any]]:
         return None
@@ -164,6 +174,9 @@ class AmazonAdsStream(HttpStream, BasicAmazonAdsStream):
             f"{response.request.headers['Amazon-Advertising-API-Scope']} profile: {resp.details}"
         )
 
+    def get_error_handler(self) -> ErrorHandler:
+        return AmazonAdsErrorHandler(logger=LOGGER)
+
 
 class SubProfilesStream(AmazonAdsStream):
     """
@@ -205,7 +218,7 @@ class SubProfilesStream(AmazonAdsStream):
         Iterate through self._profiles list and send read all records for each profile.
         """
         for profile in self._profiles:
-            self._current_profile_id = profile.profileId
+            self._current_profile_id = profile["profileId"]
             yield from super().read_records(*args, **kwargs)
 
     def request_headers(self, *args, **kwargs) -> MutableMapping[str, Any]:
