@@ -7,8 +7,7 @@ from typing import Any, List, Mapping
 
 import requests
 from airbyte_cdk.sources.streams.http.requests_native_auth import Oauth2Authenticator
-from source_amazon_ads.schemas import Profile
-from source_amazon_ads.streams.report_streams.report_stream_models import RecordType, ReportInfo
+from source_amazon_ads.streams.report_streams.report_stream_models import ReportInfo
 from source_amazon_ads.streams.report_streams.report_streams import ReportStream
 
 METRICS_MAP_V3 = {
@@ -323,7 +322,7 @@ class SponsoredDisplayReportStream(ReportStream):
     metrics_map = METRICS_MAP_V3
     metrics_type_to_id_map = METRICS_TYPE_TO_ID_MAP
 
-    def __init__(self, config: Mapping[str, Any], profiles: List[Profile], authenticator: Oauth2Authenticator):
+    def __init__(self, config: Mapping[str, Any], profiles: List[dict[str, Any]], authenticator: Oauth2Authenticator):
         super().__init__(config, profiles, authenticator)
         # using session without auth as API returns 400 bad request if Authorization header presents in request
         # X-Amz-Algorithm and X-Amz-Signature query params already present in the url, that is enough to make proper request
