@@ -19,7 +19,6 @@ from airbyte_cdk.models import SyncMode
 from airbyte_cdk.sources.streams.availability_strategy import AvailabilityStrategy
 from airbyte_cdk.sources.streams.http.requests_native_auth import Oauth2Authenticator
 from pendulum import Date
-from source_amazon_ads.schemas import CatalogModel, MetricsReport, Profile
 from source_amazon_ads.streams.common import BasicAmazonAdsStream
 from source_amazon_ads.utils import get_typed_env, iterate_one_by_one
 
@@ -80,13 +79,12 @@ class ReportStream(BasicAmazonAdsStream, ABC):
         (406, re.compile(r"^Report date is too far in the past\.")),
     ]
 
-    def __init__(self, config: Mapping[str, Any], profiles: List[Profile], authenticator: Oauth2Authenticator):
+    def __init__(self, config: Mapping[str, Any], profiles: List[dict[str, Any]], authenticator: Oauth2Authenticator):
         super().__init__(config, profiles)
         self._state = {}
         self._session = requests.Session()
         self._session.auth = authenticator
         self._report_download_session = self._session
-        self._model = self._generate_model()
         self._start_date: Optional[Date] = config.get("start_date")
         self._look_back_window: int = config["look_back_window"]
         # Timeout duration in minutes for Reports. Default is 180 minutes.
@@ -94,10 +92,6 @@ class ReportStream(BasicAmazonAdsStream, ABC):
         # Maximum retries Airbyte will attempt for fetching report data. Default is 5.
         self.report_generation_maximum_retries: int = get_typed_env("REPORT_GENERATION_MAX_RETRIES", 5)
         self._report_record_types = config.get("report_record_types")
-
-    @property
-    def model(self) -> CatalogModel:
-        return self._model
 
     @property
     def availability_strategy(self) -> Optional["AvailabilityStrategy"]:
@@ -133,13 +127,13 @@ class ReportStream(BasicAmazonAdsStream, ABC):
 
         for report_info in report_info_list:
             for metric_object in report_info.metric_objects:
-                yield self._model(
-                    profileId=report_info.profile_id,
-                    recordType=report_info.record_type,
-                    reportDate=report_date,
-                    recordId=self.get_record_id(metric_object, report_info.record_type),
-                    metric=metric_object,
-                ).dict()
+                yield {
+                    "profileId": report_info.profile_id,
+                    "recordType": report_info.record_type,
+                    "reportDate": report_date,
+                    "recordId": self.get_record_id(metric_object, report_info.record_type),
+                    "metric": metric_object,
+                }
 
     def get_record_id(self, metric_object: dict, record_type: str) -> str:
         return metric_object.get(self.metrics_type_to_id_map[record_type]) or str(uuid.uuid4())
@@ -161,7 +155,7 @@ class ReportStream(BasicAmazonAdsStream, ABC):
         return wrapped
 
     @backoff_max_tries
-    def _init_and_try_read_records(self, profile: Profile, report_date):
+    def _init_and_try_read_records(self, profile: dict[str, Any], report_date: str):
         report_info_list = self._init_reports(profile, report_date)
         self.logger.info(f"Waiting for {len(report_info_list)} report(s) to be generated")
         self._try_read_records(report_info_list)
@@ -191,17 +185,6 @@ class ReportStream(BasicAmazonAdsStream, ABC):
 
     def _incomplete_report_info(self, report_info_list):
         return [r for r in report_info_list if r.status != Status.SUCCESS and r.status != Status.COMPLETED]
-
-    def _generate_model(self):
-        """
-        Generate pydantic model based on combined list of all the metrics
-        attributes for particular stream. This model later will be used for
-        discover schema generation.
-        """
-        metrics = set()
-        for metric_list in self.metrics_map.values():
-            metrics.update(set(metric_list))
-        return MetricsReport.generate_metric_model(metrics)
 
     def _get_auth_headers(self, profile_id: int):
         return (
@@ -277,9 +260,9 @@ class ReportStream(BasicAmazonAdsStream, ABC):
             yield start_date.format(self.REPORT_DATE_FORMAT)
             start_date = start_date.add(days=1)
 
-    def get_start_date(self, profile: Profile, stream_state: Mapping[str, Any]) -> Date:
-        today = pendulum.today(tz=profile.timezone).date()
-        start_date = stream_state.get(str(profile.profileId), {}).get(self.cursor_field)
+    def get_start_date(self, profile: dict[str, Any], stream_state: Mapping[str, Any]) -> Date:
+        today = pendulum.today(tz=profile["timezone"]).date()
+        start_date = stream_state.get(str(profile["profileId"]), {}).get(self.cursor_field)
         if start_date:
             start_date = pendulum.from_format(start_date, self.REPORT_DATE_FORMAT).date()
             # Taking date from state if it's not older than 60 days
@@ -288,9 +271,9 @@ class ReportStream(BasicAmazonAdsStream, ABC):
             return max(self._start_date, today.subtract(days=self.REPORTING_PERIOD))
         return today
 
-    def stream_profile_slices(self, profile: Profile, stream_state: Mapping[str, Any]) -> Iterable[Mapping[str, Any]]:
+    def stream_profile_slices(self, profile: dict[str, Any], stream_state: Mapping[str, Any]) -> Iterable[Mapping[str, Any]]:
         start_date = self.get_start_date(profile, stream_state)
-        for report_date in self.get_date_range(start_date, profile.timezone):
+        for report_date in self.get_date_range(start_date, profile["timezone"]):
             yield {"profile": profile, self.cursor_field: report_date}
 
     def stream_slices(
@@ -319,16 +302,16 @@ class ReportStream(BasicAmazonAdsStream, ABC):
     def get_updated_state(self, current_stream_state: Dict[str, Any], latest_data: Mapping[str, Any]) -> Mapping[str, Any]:
         return self._state
 
-    def _update_state(self, profile: Profile, report_date: str):
+    def _update_state(self, profile: dict[str, Any], report_date: str):
         report_date = pendulum.from_format(report_date, self.REPORT_DATE_FORMAT).date()
-        look_back_date = pendulum.today(tz=profile.timezone).date().subtract(days=self._look_back_window - 1)
+        look_back_date = pendulum.today(tz=profile["timezone"]).date().subtract(days=self._look_back_window - 1)
         start_date = self.get_start_date(profile, self._state)
         updated_state = max(min(report_date, look_back_date), start_date).format(self.REPORT_DATE_FORMAT)
 
-        stream_state_value = self._state.get(str(profile.profileId), {}).get(self.cursor_field)
+        stream_state_value = self._state.get(str(profile["profileId"]), {}).get(self.cursor_field)
         if stream_state_value:
             updated_state = max(updated_state, stream_state_value)
-        self._state.setdefault(str(profile.profileId), {})[self.cursor_field] = updated_state
+        self._state.setdefault(str(profile["profileId"]), {})[self.cursor_field] = updated_state
 
     @abstractmethod
     def _get_init_report_body(self, report_date: str, record_type: str, profile) -> Dict[str, Any]:
@@ -341,7 +324,7 @@ class ReportStream(BasicAmazonAdsStream, ABC):
         ReportInitFailure,
         max_tries=5,
     )
-    def _init_reports(self, profile: Profile, report_date: str) -> List[ReportInfo]:
+    def _init_reports(self, profile: dict[str, Any], report_date: str) -> List[ReportInfo]:
         """
         Send report generation requests for all profiles and for all record types for specific day.
         :report_date - date for generating metric report.
@@ -362,15 +345,15 @@ class ReportStream(BasicAmazonAdsStream, ABC):
                 # different metric list for each record.
                 request_record_type = record_type.split("_")[0]
                 self.logger.info(
-                    f"Initiating report generation for {profile.profileId} profile with {record_type} type for {report_date} date"
+                    f"Initiating report generation for {profile['profileId']} profile with {record_type} type for {report_date} date"
                 )
                 response = self._send_http_request(
                     urljoin(self._url, self.report_init_endpoint(request_record_type)),
-                    profile.profileId,
+                    profile["profileId"],
                     report_init_body,
                 )
                 if response.status_code != self.report_is_created:
-                    error_msg = f"Unexpected HTTP status code {response.status_code} when registering {record_type}, {type(self).__name__} for {profile.profileId} profile: {response.text}"
+                    error_msg = f"Unexpected HTTP status code {response.status_code} when registering {record_type}, {type(self).__name__} for {profile['profileId']} profile: {response.text}"
                     if self._skip_known_errors(response):
                         self.logger.warning(error_msg)
                         break
@@ -381,7 +364,7 @@ class ReportStream(BasicAmazonAdsStream, ABC):
                     ReportInfo(
                         report_id=response.reportId,
                         record_type=record_type,
-                        profile_id=profile.profileId,
+                        profile_id=profile["profileId"],
                         status=Status.IN_PROGRESS,
                         metric_objects=[],
                     )
