@@ -2,25 +2,18 @@
 # Copyright (c) 2023 Airbyte, Inc., all rights reserved.
 #
 
-import json
+import logging
 from abc import ABC
 from http import HTTPStatus
 from typing import Any, Iterable, List, Mapping, MutableMapping, Optional
 
-from airbyte_cdk.models import SyncMode
+from airbyte_cdk.models import FailureType, SyncMode
+from airbyte_cdk.sources.streams.http.error_handlers import ErrorHandler, ErrorResolution, HttpStatusErrorHandler, ResponseAction
+from airbyte_cdk.sources.streams.http.error_handlers.default_error_mapping import DEFAULT_ERROR_MAPPING
 from requests import Response
-from source_amazon_ads.schemas import (
-    ProductAd,
-    ProductAdGroupBidRecommendations,
-    ProductAdGroups,
-    ProductAdGroupSuggestedKeywords,
-    ProductCampaign,
-    ProductTargeting,
-    SponsoredProductCampaignNegativeKeywordsModel,
-    SponsoredProductKeywordsModel,
-    SponsoredProductNegativeKeywordsModel,
-)
 from source_amazon_ads.streams.common import SubProfilesStream
+
+LOGGER = logging.getLogger("airbyte")
 
 
 class SponsoredProductsV3(SubProfilesStream):
@@ -75,7 +68,6 @@ class SponsoredProductCampaigns(SponsoredProductsV3):
     primary_key = "campaignId"
     data_field = "campaigns"
     state_filter = None
-    model = ProductCampaign
     content_type = "application/vnd.spCampaign.v3+json"
 
     def path(self, **kwargs) -> str:
@@ -99,7 +91,6 @@ class SponsoredProductAdGroups(SponsoredProductsV3):
     primary_key = "adGroupId"
     data_field = "adGroups"
     content_type = "application/vnd.spAdGroup.v3+json"
-    model = ProductAdGroups
 
     def path(self, **kwargs) -> str:
         return "/sp/adGroups/list"
@@ -128,31 +119,25 @@ class SponsoredProductAdGroupWithSlicesABC(SponsoredProductsV3, ABC):
         if response.status_code == HTTPStatus.OK:
             yield resp
 
-        if response.status_code == HTTPStatus.BAD_REQUEST:
-            # 400 error message for bids recommendation:
-            #   Bid recommendation for AD group in Manual Targeted Campaign is not supported.
-            # 400 error message for keywords recommendation:
-            #   Getting keyword recommendations for AD Group in Auto Targeted Campaign is not supported
-            self.logger.warning(
-                f"Skip current AdGroup because it does not support request {response.request.url} for "
-                f"{response.request.headers['Amazon-Advertising-API-Scope']} profile: {response.text}"
-            )
-        elif response.status_code == HTTPStatus.UNPROCESSABLE_ENTITY:
-            # 422 error message for bids recommendation:
-            # No recommendations can be provided as the input ad group does not have any asins.
-            self.logger.warning(
-                f"Skip current AdGroup because the ad group {json.loads(response.request.body)['adGroupId']} does not have any asins {response.request.url}"
-            )
-        elif response.status_code == HTTPStatus.NOT_FOUND:
-            # 404 Either the specified ad group identifier was not found,
-            # or the specified ad group was found but no associated bid was found.
-            self.logger.warning(
-                f"Skip current AdGroup because the specified ad group has no associated bid {response.request.url} for "
-                f"{response.request.headers['Amazon-Advertising-API-Scope']} profile: {response.text}"
-            )
-
-        else:
-            response.raise_for_status()
+    def get_error_handler(self) -> ErrorHandler:
+        error_mapping = DEFAULT_ERROR_MAPPING | {
+            400: ErrorResolution(
+                response_action=ResponseAction.IGNORE,
+                failure_type=FailureType.config_error,
+                error_message="Skip current AdGroup because it does not support request {response.request.url} for current profile",
+            ),
+            422: ErrorResolution(
+                response_action=ResponseAction.IGNORE,
+                failure_type=FailureType.config_error,
+                error_message="Skip current AdGroup because the ad group {json.loads(response.request.body)['adGroupId']} does not have any asins",
+            ),
+            404: ErrorResolution(
+                response_action=ResponseAction.IGNORE,
+                failure_type=FailureType.config_error,
+                error_message="Skip current AdGroup because the specified ad group has no associated bid",
+            ),
+        }
+        return HttpStatusErrorHandler(logger=LOGGER, error_mapping=error_mapping)
 
 
 class SponsoredProductAdGroupBidRecommendations(SponsoredProductAdGroupWithSlicesABC):
@@ -164,7 +149,6 @@ class SponsoredProductAdGroupBidRecommendations(SponsoredProductAdGroupWithSlice
     primary_key = None
     data_field = "bidRecommendations"
     content_type = "application/vnd.spthemebasedbidrecommendation.v4+json"
-    model = ProductAdGroupBidRecommendations
 
     def path(self, stream_slice: Mapping[str, Any] = None, **kwargs) -> str:
         return "/sp/targets/bid/recommendations"
@@ -203,7 +187,6 @@ class SponsoredProductAdGroupSuggestedKeywords(SponsoredProductAdGroupWithSlices
 
     primary_key = None
     data_field = ""
-    model = ProductAdGroupSuggestedKeywords
 
     @property
     def http_method(self, **kwargs) -> str:
@@ -238,7 +221,6 @@ class SponsoredProductKeywords(SponsoredProductsV3):
     primary_key = "keywordId"
     data_field = "keywords"
     content_type = "application/vnd.spKeyword.v3+json"
-    model = SponsoredProductKeywordsModel
 
     def path(self, **kwargs) -> str:
         return "sp/keywords/list"
@@ -253,7 +235,6 @@ class SponsoredProductNegativeKeywords(SponsoredProductsV3):
     primary_key = "keywordId"
     data_field = "negativeKeywords"
     content_type = "application/vnd.spNegativeKeyword.v3+json"
-    model = SponsoredProductNegativeKeywordsModel
 
     def path(self, **kwargs) -> str:
         return "sp/negativeKeywords/list"
@@ -268,7 +249,6 @@ class SponsoredProductCampaignNegativeKeywords(SponsoredProductsV3):
     primary_key = "keywordId"
     data_field = "campaignNegativeKeywords"
     content_type = "application/vnd.spCampaignNegativeKeyword.v3+json"
-    model = SponsoredProductCampaignNegativeKeywordsModel
 
     def path(self, **kwargs) -> str:
         return "sp/campaignNegativeKeywords/list"
@@ -283,7 +263,6 @@ class SponsoredProductAds(SponsoredProductsV3):
     primary_key = "adId"
     data_field = "productAds"
     content_type = "application/vnd.spProductAd.v3+json"
-    model = ProductAd
 
     def path(self, **kwargs) -> str:
         return "sp/productAds/list"
@@ -297,7 +276,6 @@ class SponsoredProductTargetings(SponsoredProductsV3):
     primary_key = "targetId"
     data_field = "targetingClauses"
     content_type = "application/vnd.spTargetingClause.v3+json"
-    model = ProductTargeting
 
     def path(self, **kwargs) -> str:
         return "sp/targets/list"
