@@ -9,7 +9,6 @@ from airbyte_cdk.models import SyncMode
 from airbyte_cdk.test.catalog_builder import CatalogBuilder
 from airbyte_cdk.test.entrypoint_wrapper import EntrypointOutput, read
 from source_amazon_ads import SourceAmazonAds
-from source_amazon_ads.declarative_source_adapter import DeclarativeSourceAdapter
 
 
 def read_stream(
@@ -20,7 +19,7 @@ def read_stream(
     expecting_exception: bool = False
 ) -> EntrypointOutput:
     catalog = CatalogBuilder().with_stream(stream_name, sync_mode).build()
-    return read(DeclarativeSourceAdapter(source=SourceAmazonAds()), config, catalog, state, expecting_exception)
+    return read(SourceAmazonAds(), config, catalog, state, expecting_exception)
 
 
 def get_log_messages_by_log_level(logs: List[AirbyteMessage], log_level: LogLevel) -> List[str]:
