@@ -7,8 +7,6 @@ import responses
 from airbyte_cdk.models import AirbyteConnectionStatus, AirbyteMessage, ConnectorSpecification, Status, Type
 from jsonschema import Draft4Validator
 from source_amazon_ads import SourceAmazonAds
-from source_amazon_ads.declarative_source_adapter import DeclarativeSourceAdapter
-from source_amazon_ads.schemas import Profile
 
 from .utils import command_check, url_strip_query
 
@@ -41,17 +39,17 @@ def ensure_additional_property_is_boolean(root):
 @responses.activate
 def test_discover(config):
     setup_responses()
-    source = DeclarativeSourceAdapter(source=SourceAmazonAds())
+    source = SourceAmazonAds()
     catalog = source.discover(None, config)
-    catalog = AirbyteMessage(type=Type.CATALOG, catalog=catalog).dict(exclude_unset=True)
-    schemas = [stream["json_schema"] for stream in catalog["catalog"]["streams"]]
+    catalog = AirbyteMessage(type=Type.CATALOG, catalog=catalog)
+    schemas = [stream.json_schema for stream in catalog.catalog.streams]
     for schema in schemas:
         Draft4Validator.check_schema(schema)
         ensure_additional_property_is_boolean(schema)
 
 
 def test_spec():
-    source = DeclarativeSourceAdapter(source=SourceAmazonAds())
+    source = SourceAmazonAds()
     spec = source.spec(None)
     assert isinstance(spec, ConnectorSpecification)
 
@@ -59,7 +57,7 @@ def test_spec():
 @responses.activate
 def test_check(config_gen):
     setup_responses()
-    source = DeclarativeSourceAdapter(source=SourceAmazonAds())
+    source = SourceAmazonAds()
 
     assert command_check(source, config_gen(start_date=...)) == AirbyteConnectionStatus(status=Status.SUCCEEDED)
     assert len(responses.calls) == 2
@@ -90,7 +88,7 @@ def test_check(config_gen):
 @responses.activate
 def test_source_streams(config):
     setup_responses()
-    source = DeclarativeSourceAdapter(source=SourceAmazonAds())
+    source = SourceAmazonAds()
     streams = source.streams(config)
     assert len(streams) == 27
     actual_stream_names = {stream.name for stream in streams}
@@ -129,14 +127,14 @@ def test_filter_profiles_exist():
         {"profileId": 333, "timezone": "gtm", "accountInfo": {"marketplaceStringId": "mkt_id_3", "id": "333", "type": "vendor"}},
     ]
 
-    mock_profiles = [Profile.parse_obj(profile) for profile in mock_objs]
+    mock_profiles = [profile for profile in mock_objs]
 
     filtered_profiles = source._choose_profiles({}, mock_profiles)
     assert len(filtered_profiles) == 3
 
     filtered_profiles = source._choose_profiles({"profiles": [111]}, mock_profiles)
     assert len(filtered_profiles) == 1
-    assert filtered_profiles[0].profileId == 111
+    assert filtered_profiles[0]['profileId'] == 111
 
     filtered_profiles = source._choose_profiles({"profiles": [111, 333]}, mock_profiles)
     assert len(filtered_profiles) == 2
@@ -149,7 +147,7 @@ def test_filter_profiles_exist():
 
     filtered_profiles = source._choose_profiles({"marketplace_ids": ["mkt_id_1"]}, mock_profiles)
     assert len(filtered_profiles) == 1
-    assert filtered_profiles[0].accountInfo.marketplaceStringId == "mkt_id_1"
+    assert filtered_profiles[0]['accountInfo']['marketplaceStringId'] == "mkt_id_1"
 
     filtered_profiles = source._choose_profiles({"marketplace_ids": ["mkt_id_1", "mkt_id_3"]}, mock_profiles)
     assert len(filtered_profiles) == 2
@@ -159,4 +157,4 @@ def test_filter_profiles_exist():
 
     filtered_profiles = source._choose_profiles({"profiles": [111], "marketplace_ids": ["mkt_id_1"]}, mock_profiles)
     assert len(filtered_profiles) == 1
-    assert filtered_profiles[0].profileId == 111
+    assert filtered_profiles[0]['profileId'] == 111
